policies_gates//Event/Auth/Edit_Update_Delete/

+++ app/Http/Controllers/EventController.php
@@ -4,9 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
+    public function __construct(){
+        $this->middleware('guest'); // unfinished
+    }
     /**
      * Display a listing of the resource.
      *
@@ -77,6 +81,9 @@ class EventController extends Controller
      */
     public function edit(Event $event)
     {
+        if(! Gate::allows('only-admin')){
+            abort(403);
+        }
         return view('events.edit', compact('event'));
     }
 
@@ -89,6 +96,8 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+        $this->authorize('update', $event);
+
         $event->update([
             'name' => $request['name'],
         ]);
@@ -132,6 +141,15 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        /*
+        if(! Gate::allows('only-admin')){
+            abort(403);
+        }
+ */
+        $this->authorize('delete', $event);
+
+        $this->deleteAllUserInEvent($event);
+
         $event->delete();
         return redirect()->route('events.index');
     }
